atualizando header no app.js
