+++ src/components/Vender/Vender.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import styled from 'styled-components'
+
+
+export default class Vender extends React.Component {
+    render(){
+        return(
+            <div>Vender</div>
+        );
+    }
+}
